[Spark] Split UCProxySuite: run table/namespace tests on 4.0/4.1/4.2
yili-db noted the moved UCProxySuite ran only on Spark 4.2, leaving the
version-agnostic table/namespace surface untested on 4.0/4.1. Split it:

- New UCProxyTestFixture (src/test/java, non-test helper): the reflective
  UCProxy construction + mock injection, exposed via composition so neither
  suite inherits @Test methods (which JUnit5 would otherwise double-run on 4.2).
  Holds the proxy as a bare Object so it names no Spark-4.2-only type.
- UCProxySuite moves to src/test/java and keeps the 10 version-agnostic tests
  (listTables x5, listNamespaces x3, loadTable single-RPC guard, dropTable
  metric-view safety). It now compiles and runs on 4.0, 4.1, and 4.2 -- the
  coverage that was previously 4.2-only.
- UCViewProxySuite (src/test/scala-shims/spark-4.2, was UCProxySuite) keeps the
  9 view/4.2-only tests (createView/loadView/dropView/loadTableOrView, the V2
  TableInfo createTable overload, MetadataTable/ViewInfo assertions). It casts
  the proxy to TableViewCatalog/ViewCatalog, which only succeeds on 4.2.

Also makes testLoadTableTriggersExactlyOneGetTableRpc hermetic: loadTable issues
its single getTable RPC before credential vending (which has no real endpoint in
the mock context), so the call is guarded and only the RPC count is asserted.
This both fixes a pre-existing 4.2 failure and makes the test safe to run on all
versions.

Verified: 4.2 = 19/19 (10 shared + 9 view, each once), 4.0 = 10/10. 4.1 shares
the same source as 4.0/4.2 and is exercised in CI.
